+++ esp-storage/CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- The `storage` and `nor-flash` features have been removed, the related functionality is now always available. (#3431)
+
 ## [0.5.0] - 2025-02-24
 
 ### Changed