+++ esp-storage/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Bump Rust edition to 2024, bump MSRV to 1.85. (#3391)
+- The `low-level` feature has been removed, the gated API is always available (#3425)
 
 ### Fixed